requirements.txt
Adafruit-CharLCD==1.1.1
Adafruit-DHT==1.4.0
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.3
certifi==2018.8.24
RPi.GPIO==0.6.5
spidev==3.4
